Stanford Prospectus files
These documents were created when we were soliciting other groups on campus to join with us in a Carpentries partnership. PDF provided for easy access. Original created in Adobe InDesign, which is also included. 

The photograph is by me. Please credit me if it doesn't completely screw up your graphic design. Those photographed granted permission for use of the images by Stanford Libraries and the Carpentries for print and web materials. I can dig up the exact language if anyone wants to use the image to make sure the use complies with the permissions.