Instructor Training Survey
This survey was used to help decide which of the people who had expressed interest in instructor training as part of our partnership agreement would be offered the limited number of seats at the training. This was created in Google Forms and should be viewable at https://docs.google.com/forms/d/1z5U2v4APcTwvuegCM8yAjgwUtUyHYiaEnMWrQkPSsh0/edit in its original form. Recreated here as a pdf for better access.